fix: handle specific bad request
